:rocket: init refactoring
